+++ comics/comics/atheistcartoons.py
@@ -6,19 +6,10 @@ class ComicData(ComicDataBase):
     language = 'en'
     url = 'http://www.atheistcartoons.com/'
     start_date = '2009-01-03'
+    end_date = '2011-08-25'
+    active = False
     rights = 'Atheist Cartoons'
 
 class Crawler(CrawlerBase):
-    history_capable_date = '2009-01-03'
-    schedule = None
-    time_zone = 9
-
     def crawl(self, pub_date):
-        page_url = 'http://www.atheistcartoons.com/?m=%s' % \
-            pub_date.strftime('%Y%m%d')
-        page = self.parse_page(page_url)
-
-        url = page.src('div[class="entry"] img')
-        title = page.text('div[class="entry"] > h2 a')
-
-        return CrawlerImage(url, title)
+        pass # Comic no longer published